+++ Examples/Console/DIO/example_DIO_loopback_pins.py
@@ -13,7 +13,7 @@ async def main():
     ## Create device handle
     dev = pywpc.USBDAQF1D()
 
-    ## Connect to network device
+    ## Connect to USB device
     try:
         dev.connect('21JA1044')
     except Exception as err:
@@ -32,9 +32,9 @@ async def main():
         status = await dev.DO_openPins(port, [0,1,2,3,4]) 
         if status == 0: print("DO_openPins: OK")
         
-        ## Set pin0, pin1 to high, others to low.
-        await dev.DO_writeValuePins(port, [0,1,2,3,4], [1,1,0,0,0]) 
-        if status == 0: print("DO_writeValuePins: OK")
+        ## Set pin0, pin1 to high, others to low
+        await dev.DO_writePins(port, [0,1,2,3,4], [1,1,0,0,0]) 
+        if status == 0: print("DO_writePins: OK")
 
         ## Open pin5, pin6 and pin7 in port 0 with digital output
         status = await dev.DI_openPins(port, [5,6,7])
@@ -57,7 +57,7 @@ async def main():
     except Exception as err:
         pywpc.printGenericError(err)
 
-    ## Disconnect network device
+    ## Disconnect USB device
     dev.disconnect()
     
     ## Release device handle