+++ Examples/Console/DIO/example_DIO_loopback_port.py
@@ -13,7 +13,7 @@ async def main():
     ## Create device handle
     dev = pywpc.USBDAQF1D()
 
-    ## Connect to network device
+    ## Connect to USB device
     try:
         dev.connect('21JA1044')
     except Exception as err:
@@ -34,8 +34,8 @@ async def main():
         if status == 0: print("DO_openPort in port 0: OK")
 
         ## Set pin0, pin1 and pin2 to high, others to low
-        status = await dev.DO_writeValuePort(port_DO, [0,0,0,0,0,1,1,1])
-        if status == 0: print("DO_writeValuePort: OK")
+        status = await dev.DO_writePort(port_DO, [0,0,0,0,0,1,1,1])
+        if status == 0: print("DO_writePort: OK")
 
         ## Open all pins in port 1 with digital input
         status = await dev.DI_openPort(port_DI)
@@ -58,7 +58,7 @@ async def main():
     except Exception as err:
         pywpc.printGenericError(err)
         
-    ## Disconnect network device
+    ## Disconnect USB device
     dev.disconnect()
 
     ## Release device handle