+++ Examples/Console/Thermo/example_Thermo_read_channel_data.py
@@ -13,7 +13,7 @@ async def main():
     ## Create device handle
     dev = pywpc.USBDAQF1TD()
 
-    ## Connect to network device
+    ## Connect to USB device
     try:
         dev.connect('21JA1239')
     except Exception as err:
@@ -31,39 +31,39 @@ async def main():
         over_sampling_mode = 0  ## 0:1 sample, 1:2 samples, 2:4 sample, 3:8 samples, 4:16 samples
         thermo_type = 3         ## 0:B type, 1:E type, 2:J type, 3:K type
                                 ## 4:N type, 5:R type, 6:S type, 7:T type
-   
-        ## Open thermo
+        
+        ## Open thermo port1
         status = await dev.Thermal_open(port)
         if status == 0: print("Thermal_open: OK")
 
         ## Sleep
         await asyncio.sleep(0.1) ## delay(second)
         
-        ## Set over-sampling mode to no over-sampling in channel 1 
+        ## Set thermo port to 1 and set over-sampling mode to no over-sampling in channel 1 
         status = await dev.Thermal_setOverSampling(port, channel, over_sampling_mode)
         if status == 0: print("setOverSampling: OK")   
 
         ## Sleep
         await asyncio.sleep(0.1) ## delay(second)
         
-        ## Set K type in channel 1 
+        ## Set thermo port to 1 and set K type in channel 1 
         status = await dev.Thermal_setType(port, channel, thermo_type)
         if status == 0: print("setType: OK")   
  
         ## Sleep
         await asyncio.sleep(0.1) ## delay(second)
 
-        ## Read thermo in channel 1
+        ## Set thermo port to 1 and read thermo in channel 1
         data = await dev.Thermal_readSensor(port, channel)
         print("Read channel 1 data:", data, "°C")
 
-        ## Close thermo
+        ## Close thermo port1
         status = await dev.Thermal_close(port)
         if status == 0: print("Thermal_close: OK")   
     except Exception as err:
         pywpc.printGenericError(err)
 
-    ## Disconnect network device
+    ## Disconnect USB device
     dev.disconnect()
 
     ## Release device handle