+++ Examples/GUI/System/Example_find_all_device/main.py
@@ -21,7 +21,7 @@ class MainWindow(QtWidgets.QMainWindow):
         self.ui = Ui_MainWindow()
         self.ui.setupUi(self)
 
-        ## Trademark path
+        ## Set trademark path
         trademark_path = os.getcwd() + "\Material\WPC_trademark.jpg" 
         self.ui.lb_trademark.setPixmap(QtGui.QPixmap(trademark_path))
 
@@ -33,11 +33,10 @@ class MainWindow(QtWidgets.QMainWindow):
         
         ## Connect to network device
         dev.connect() 
+ 
+        ## Get Python driver version 
+        print(f'{pywpc.PKG_FULL_NAME} - Version {pywpc.__version__}')
 
-        ## Get Python driver version
-        str_ = f'{pywpc.PKG_FULL_NAME} - Version {pywpc.__version__}'
-        print(str_)
-  
         ## Initialize list
         self.broadcast_list=[]
 
@@ -47,7 +46,7 @@ class MainWindow(QtWidgets.QMainWindow):
         self.ui.tableWidget_brst.setStyleSheet("selection-background-color: #217536;")
 
     def loaddata(self): 
-        row=0
+        row = 0
         self.ui.tableWidget_brst.setRowCount(len(self.broadcast_list))
         for i in self.broadcast_list:
             self.ui.tableWidget_brst.setItem(row, 0, QtWidgets.QTableWidgetItem(i["ip"]))
@@ -55,23 +54,31 @@ class MainWindow(QtWidgets.QMainWindow):
             self.ui.tableWidget_brst.setItem(row, 2, QtWidgets.QTableWidgetItem(i["mac"]))
             self.ui.tableWidget_brst.setItem(row, 3, QtWidgets.QTableWidgetItem(i["model"]))
             self.ui.tableWidget_brst.setItem(row, 4, QtWidgets.QTableWidgetItem(i["firmware_ver"]))
-            row=row+1
+            row = row + 1
         QApplication.restoreOverrideCursor()
         
     @asyncSlot()      
     async def broadcastEvent(self):
+
+        ## Add WaitCursor
         QApplication.setOverrideCursor(Qt.WaitCursor)
+
+        ## Clear Broadcast table
         self.broadcast_list.clear()
         self.ui.tableWidget_brst.setRowCount(0) 
-        broadcast_info = await dev.brst_getDeviceInfo()
+
+        ## Perform device information
+        broadcast_info = await dev.Bcst_getDeviceInfo()
+
+        ## Return information length
         device_amount = len(broadcast_info) 
         for i in range (device_amount):
-            ip      = broadcast_info[i][0]
+            ip = broadcast_info[i][0]
             submask = broadcast_info[i][1]
-            mac     = broadcast_info[i][2]
+            mac = broadcast_info[i][2]
             model_version = broadcast_info[i][3]
-            str_list      = model_version.split('_')
-            model   = str_list[0]
+            str_list = model_version.split('_')
+            model = str_list[0]
             version = str_list[1] 
             self.broadcast_list.append({'ip': ip , 'submask': submask, 'mac':mac, 'model': model, 'firmware_ver': version}) 
         self.loaddata()
@@ -79,6 +86,7 @@ class MainWindow(QtWidgets.QMainWindow):
     def closeEvent(self, event):
         ## Disconnect network device
         dev.disconnect()
+        
         ## Release device handle
         dev.close()
 
@@ -92,7 +100,7 @@ def main():
         loop.run_forever()
 
 if __name__ == "__main__":
-    ## Create device handle
+    ## Create device handle 
     dev = pywpc.Broadcaster()
     main()